Preliminary fix for getting the title
Now the userscript will send a message with the data to the background script.
The background script will decide what to do with it.

The userscript is started at the document start, so as soon as we have something
 the background script will close the tab.

Now we handle closing the tab as well.

TODO:

 - Let the user define the max time to look for information
 - Ignore favicon if the page is completely loaded and really doesn't
   contain one

LoveIsGrief/links-for-later#31 Store also the title of the linked page